requirements.txt
langchain
julep
pydantic
requests
pyyaml
langchain-core